clewan up base
